flask_backend/requirements.txt
Flask
Flask-CORS
pandas
openpyxl
PyJWT
cryptography
python-dotenv
werkzeug
bcrypt